+++ plugins/ocm-backend/src/types.ts
@@ -1,4 +1,5 @@
 import { KubernetesObject, V1PodCondition } from '@kubernetes/client-node';
+import { TaskScheduleDefinition } from '@backstage/backend-tasks';
 
 export type OcmConfig = {
   id: string;
@@ -8,6 +9,7 @@ export type OcmConfig = {
   skipTLSVerify?: boolean;
   caData?: string;
   owner: string;
+  schedule?: TaskScheduleDefinition;
 };
 
 export interface ClusterClaim {
@@ -47,7 +49,7 @@ export interface ManagedClusterInfo extends KubernetesObject {
     masterEndpoint: string;
   };
   status?: {
-    nodeList: {
+    nodeList?: {
       capacity: {
         cpu: string;
         memory: string;