+++ plugins/ocm/src/components/ClusterStatusPage/ClusterStatusPage.tsx
@@ -1,16 +1,17 @@
-import React, { useState } from 'react';
+import React, { ReactElement, useState } from 'react';
 import { SearchContextProvider } from '@backstage/plugin-search-react';
 import {
   Content,
   Page,
-  InfoCard,
   WarningPanel,
   CodeSnippet,
+  Header,
+  Table,
   StatusOK,
   StatusError,
-  Header,
+  StatusAborted,
 } from '@backstage/core-components';
-import { CircularProgress, Grid, makeStyles } from '@material-ui/core';
+import { Chip, CircularProgress, Grid, makeStyles } from '@material-ui/core';
 
 import { catalogApiRef, EntityRefLink } from '@backstage/plugin-catalog-react';
 import useAsyncFn from 'react-use/lib/useAsyncFn';
@@ -19,13 +20,23 @@ import { useApi } from '@backstage/core-plugin-api';
 import { Entity } from '@backstage/catalog-model';
 import { HomePageCompanyLogo } from '@backstage/plugin-home';
 import { ErrorResponseBody } from '@backstage/errors';
-import { ClusterOverview } from '@janus-idp/backstage-plugin-ocm-common';
+import {
+  ClusterNodesStatus,
+  ClusterOverview,
+} from '@janus-idp/backstage-plugin-ocm-common';
 import { OcmApiRef } from '../../api';
+import { Status, Update } from '../common';
 
 interface clusterEntity {
-  status: boolean;
+  cluster: ClusterOverview;
   entity: Entity;
 }
+
+const useStylesTwo = makeStyles({
+  container: {
+    width: '100%',
+  },
+});
 
 const useStyles = makeStyles(theme => ({
   container: {
@@ -37,38 +48,59 @@ const useStyles = makeStyles(theme => ({
   },
 }));
 
-const useCatalogStyles = makeStyles({
-  root: {
-    height: '100%',
-    transition: 'all .25s linear',
-    textAlign: 'center',
-    '&:hover': {
-      boxShadow: '0px 0px 16px 0px rgba(0,0,0,0.8)',
-    },
-    '& svg': {
-      fontSize: 80,
-    },
-    '& img': {
-      height: 80,
-      width: 80,
-      objectFit: 'contain',
-    },
-  },
-  subheader: {
-    display: 'block',
-    width: '100%',
-  },
-  link: {
-    '&:hover': {
-      textDecoration: 'none',
-    },
-  },
-});
+const NodeChip = ({
+  count,
+  indicator,
+}: {
+  count: number;
+  indicator: ReactElement;
+}) => (
+  <>
+    {count > 0 ? (
+      <>
+        <Chip
+          label={
+            <>
+              {indicator}
+              {count}
+            </>
+          }
+          variant="outlined"
+        />
+      </>
+    ) : (
+      <></>
+    )}
+  </>
+);
+
+const NodeChips = ({ nodes }: { nodes: ClusterNodesStatus[] }) => {
+  const readyChipsNodes = nodes.filter(node => node.status === 'True').length;
+  // TODO: Check if not ready correctly
+  const notReadyNodesCount = nodes.filter(
+    node => node.status === 'False',
+  ).length;
+
+  if (nodes.length === 0) {
+    return <>-</>;
+  }
+
+  return (
+    <>
+      <NodeChip count={readyChipsNodes} indicator={<StatusOK />} />
+      <NodeChip count={notReadyNodesCount} indicator={<StatusError />} />
+      <NodeChip
+        count={nodes.length - readyChipsNodes - notReadyNodesCount}
+        indicator={<StatusAborted />}
+      />
+    </>
+  );
+};
 
 const CatalogClusters = () => {
   const catalogApi = useApi(catalogApiRef);
   const ocmApi = useApi(OcmApiRef);
-  const classes = useCatalogStyles();
+  const classes = useStylesTwo();
 
   const [clusterEntities, setClusterEntities] = useState<clusterEntity[]>([]);
   const [{ loading, error }, refresh] = useAsyncFn(
@@ -89,7 +121,7 @@ const CatalogClusters = () => {
             cd => cd.name === entity.metadata.name,
           );
           return {
-            status: cluster?.status.available!,
+            cluster: cluster!,
             entity: entity,
           };
         }),
@@ -112,42 +144,56 @@ const CatalogClusters = () => {
     return <CircularProgress />;
   }
 
+  const data = clusterEntities.map(ce => {
+    return {
+      name: (
+        <EntityRefLink entityRef={ce.entity}>{ce.cluster.name}</EntityRefLink>
+      ),
+      status: <Status status={ce.cluster.status} />,
+      infrastructure: ce.cluster.platform,
+      version: (
+        <Update
+          data={{
+            version: ce.cluster.openshiftVersion,
+            update: ce.cluster.update,
+          }}
+        />
+      ),
+      nodes: <NodeChips nodes={ce.cluster.nodes} />,
+    };
+  });
+
   return (
-    <>
-      {clusterEntities.map(clusterEntity => (
-        <Grid
-          item
-          xs={12}
-          sm={6}
-          md={4}
-          lg={3}
-          xl={2}
-          key={clusterEntity.entity.metadata.name}
-        >
-          <EntityRefLink
-            entityRef={clusterEntity.entity}
-            className={classes.link}
-          >
-            <InfoCard
-              divider={false}
-              noPadding
-              className={classes.root}
-              title={
-                <div className={classes.subheader}>
-                  {clusterEntity.status === true ? (
-                    <StatusOK />
-                  ) : (
-                    <StatusError />
-                  )}
-                  {clusterEntity.entity.metadata.title ||
-                    clusterEntity.entity.metadata.name}
-                </div>
-              }
-            />
-          </EntityRefLink>
-        </Grid>
-      ))}
-    </>
+    <div className={classes.container}>
+      <Table
+        options={{ paging: false }}
+        data={data}
+        columns={[
+          {
+            title: 'Name',
+            field: 'name',
+            highlight: true,
+          },
+          {
+            title: 'Status',
+            field: 'status',
+          },
+          {
+            title: 'Infrastructure',
+            field: 'infrastructure',
+          },
+          {
+            title: 'Version',
+            field: 'version',
+          },
+          {
+            title: 'Nodes',
+            field: 'nodes',
+          },
+        ]}
+        title="All"
+      />
+    </div>
   );
 };
 
@@ -160,7 +206,7 @@ export const ClusterStatusPage = ({ logo }: { logo?: React.ReactNode }) => {
         <Header title="Your Managed Clusters" />
         <Content>
           <Grid container justifyContent="center" spacing={6}>
-            <HomePageCompanyLogo className={container} logo={logo} />
+            {logo && <HomePageCompanyLogo className={container} logo={logo} />}
             <Grid container item xs={12} justifyContent="center">
               <CatalogClusters />
             </Grid>