+++ plugins/keycloak-backend/CHANGELOG.md
@@ -1,3 +1,7 @@
+## @janus-idp/backstage-plugin-keycloak-backend [1.3.7](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-keycloak-backend@1.3.6...@janus-idp/backstage-plugin-keycloak-backend@1.3.7) (2023-05-30)
+
+## @janus-idp/backstage-plugin-keycloak-backend [1.3.6](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-keycloak-backend@1.3.5...@janus-idp/backstage-plugin-keycloak-backend@1.3.6) (2023-05-30)
+
 ## @janus-idp/backstage-plugin-keycloak-backend [1.3.5](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-keycloak-backend@1.3.4...@janus-idp/backstage-plugin-keycloak-backend@1.3.5) (2023-05-22)